query multiple RPCs
